Remove TLS client refactor
